add useCustomerProductList
